Provide JSONDisk with diskcache
