+++ diskcache/core.py
@@ -7,6 +7,7 @@ import contextlib as cl
 import errno
 import functools as ft
 import io
+import json
 import os
 import os.path as op
 import pickletools
@@ -359,6 +360,34 @@ class Disk(object):
                 # ENOENT may occur if two caches attempt to delete the same
                 # file at the same time.
                 raise
+
+
+class JSONDisk(Disk):
+    """Cache key and value (de)serialized as JSON."""
+    def __init__(self, directory, compress_level=1, **kwargs):
+        self.compress_level = compress_level
+        super(JSONDisk, self).__init__(directory, **kwargs)
+
+    def put(self, key):
+        json_bytes = json.dumps(key).encode('utf-8')
+        data = zlib.compress(json_bytes, self.compress_level)
+        return super(JSONDisk, self).put(data)
+
+    def get(self, key, raw):
+        data = super(JSONDisk, self).get(key, raw)
+        return json.loads(zlib.decompress(data).decode('utf-8'))
+
+    def store(self, value, read, key=UNKNOWN):
+        if not read:
+            json_bytes = json.dumps(value).encode('utf-8')
+            value = zlib.compress(json_bytes, self.compress_level)
+        return super(JSONDisk, self).store(value, read, key=key)
+
+    def fetch(self, mode, filename, value, read):
+        data = super(JSONDisk, self).fetch(mode, filename, value, read)
+        if not read:
+            data = json.loads(zlib.decompress(data).decode('utf-8'))
+        return data
 
 
 class Timeout(Exception):
